DINF-94: mark last build as built on master to fix workspace deletion
When a build is run on e.g. an ephemeral docker node, the node will not
be available when the project is shelved.
AbstractProject.doDoWipeoutWorkspace() will then do nothing, leaving the
workspace behind and wasting disk space (or causing manual cleanup).

The workaround is to explicitly mark the last build as having been run
locally on the jenkins master, so that the master can agree to wipe the
workspace.

This will of course only work if master and slave both share the
workspace directory and this is why we can't easily upstream this.